+++ chrome-remote-cache/src/index.ts
@@ -1,6 +1,8 @@
 export { ChromeRemoteCache } from './ChromeRemoteCache';
 export { CacheStat } from './CacheStat';
 export { CacheManager } from './CacheManager';
+export { CacheManagerRedisLongTerm } from './CacheManagerRedisLongTerm';
+export { CacheManagerRedisTTL } from './CacheManagerRedisTTL';
 export { dropQueryParam, formatSize } from './CacheUtils';
 
 export { ChromeRemoteCache as default } from './ChromeRemoteCache';